fix bot message in CH

+++ chat_hub/core/domain/entities/bot_message_entity.go
@@ -4,10 +4,10 @@ import "time"
 
 type BotMessageEntity struct {
 	ID            string    `db:"id"`
-	UserId        float64   `db:"user_id"`
+	UserId        int64     `db:"user_id"`
 	DialogueId    string    `db:"dialogue_id"`
 	UserMessageId string    `db:"user_message_id"`
-	OrderNumber   float64   `db:"order_number"`
+	OrderNumber   int64     `db:"order_number"`
 	MessageTyoe   string    `db:"message_type"`
 	Content       string    `db:"content"`
 	Metadata      string    `db:"metadata"`
